Change structure of dotfiles
Now the dotfiles are in the folder home. You can add a new dotfile with
add.sh and you can add the files in the home folder with setup.sh. Of
course files are not copied but hard link, so that when change is done
in your $HOME it will also be change in home.